requirements.txt
click~=7.1.2
PyYAML==6.0
psycopg2~=2.8.6
packaging==21.3